Polishing, such as removing imports and what not
