--- core/src/main/java/com/example/demo/core/stock/StockFacade.java
@@ -1,22 +1,0 @@
-package com.example.demo.core.stock;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-@RequiredArgsConstructor
-public class StockFacade {
-
-  private final Map<Integer, StockCore> implementations = new HashMap<>();
-
-  public StockCore get(Integer version){
-    return implementations.get(version);
-  }
-
-  public void register(Integer version, StockCore stockImplementation) {
-    implementations.put(version, stockImplementation);
-  }
-}
